+++ modules/core/src/matmul.cpp
@@ -1992,7 +1992,7 @@ void cv::transform( InputArray _src, OutputArray _dst, InputArray _mtx )
         {
             for( j = 0; isDiag && j < scn; j++ )
             {
-                double v = mtype == CV_32F ? m.at<float>(i, j) : m.at<double>(i, j);
+                double v = mtype == CV_32F ? static_cast<double>(m.at<float>(i, j)) : m.at<double>(i, j);
                 if( i != j && fabs(v) > eps )
                     isDiag = false;
             }
